Build webauthn--1.0.sql with the makefile
We do not want to have to check in generated code every time we modify
something so it is better to make the build of webauthn--1.0.sql part of
the job of the makefile. We lose some code comments but I think
readability is still improved.

Arguably comaplin_header.sql does not belong under FUNCTIONS but let's
put it there for now.