Test as much output as possible from init_credential() and get_credentials()
Before we ignored the output since it contained the random challenge value,
which cannot be tested since it's random.

Now, instead, replace the challenge in the output with null,
allowing the rest of the json to be tested.

Also, use jsonb_pretty() to canonicalize the json so it can be compared.