Change verify_assertion() to only return assertion_id instead of boolean
Also drop the assertions.verified boolean column,
since all rows in the "assertions" are now guaranteed to have been verified,
otherwise no row would ever have been written.

This is to eliminate the risk a user would not check the old "verified"
to be TRUE. It's more safe to get rid of the column altogether,
and to only store verified assertions.

Returning the assertion_id also allows for integration
with other parts of an application, where a reference to the assertion
needs to be saved in some other table, to be able to lookup
what credential the user signed up or signed in with.