Move hard-coded logical checks from WHERE clause in function code to named CHECK constraints in table definitions
When a logical expression that must always hold true is part of a WHERE clause,
if it would every be false, the entire WHERE clause will fail and no row
will be returned by the SELECT. This is undesirable in our case, since we
want to be able to distinguish between different types of errors.

By moving the checks from being hard-coded in the functions source code,
and instead creating named CHECK constraints in the table definitions,
such checks are then made upon the INSERT to the table,
and will throw an error indicating which check constraint that failed.

Also:

* Properly handle credential/assertion timeouts
* Allow all timestamps to be set via input parameters to allow writing deterministis tests
* Give timestamps more precise names, instead of mostly "created_at", use "challenge_at", "credential_at", to avoid picking the wrong timestamp column.
* Fully-qualify columns/input parameters in make_credential()
* Add test make_credential_timeout
* Add test verify_assertion_timeout
* Update all tests to expect "violates check constraint" instead of expecting NULL values