Rename all tests to names that start with "error_" or "ok_"
This is to make it easier to get an overview of what is being tested.

Also rename a few CHECK constraints to more clearly express
what is being assumed will always be TRUE.

For instance, "user_verified_or_not_required" instead of just "user_verification".
This new name is hopefully all that is needed to understand
the semantics of the check constraint.