Implement setting attestation conveyance preference
This adds the ENUM "webauthn.attestation_conveyance_preference"
from the WebAuthn specification:

https://www.w3.org/TR/webauthn-2/#enum-attestation-convey

This ENUM is the type for the new column: credential_challenges.attestation

This also adds a new "attestation" input parameter to init_credential()
with the DEFAULT value 'none', adhering to the WebAuthn specification.

This is just a preparatory commit. The attestation statement is not yet parsed.