Migrate to 🧬🐘cbor for CBOR to JSON decoding of COSE public keys and attestation objects
The 🧬🐘cbor project is hosted at: https://github.com/truthly/pg-cbor,
and is a spin-off from this project, to make the functionality available
to other PostgreSQL projects in need of a CBOR to JSON decoder.

The code was not just moved, but built upon to support all cases
possible to directly translate to JSON. A rigorous test suite
was also created.