Allow controlling require_resident_key via input parameter, previously hard-coded to FALSE
Setting require_resident_key to TRUE tells the Authenticator device it must store
the *user.id* value from webauthn.init_credential() and later set user_handle to this value
when webauthn.verify_assertion() called during login.

This allows for a username-less sign-in, as the user after having signed-up with a username,
will not have to enter any username when logging in.

This concept is know as "Discoverable Credentials", and also affects webauthn.get_credentials()
which should then be called without any user_name input parameter, which will cause it to
store the input challenge like normal, but the returned allowCredentials array will be empty,
possibly thanks to the Authenticator knows what credentials are possible to login with
at the relying party's effective domain name.