Move code generating credential creation options to own function
The new function get_credential_creation_options() is useful
to add a new key to a different device, as the challenge
will then be sent via some media to the other device,
and from it, the challenge is used as input to this new function,
to get the credential creation options, needed to create
a new credential for the existing user on its other device.

This also adds the new column require_resident_key
to the table credential_challenges table, which had been forgotten,
which now caused a problem since it couldn't be saved and
then retrived from the table by get_credential_creation_options(),
as the column didn't exist.